1. move consultation question response stuff from set ups to registry. 2. correct routing bugs in consultation question stuff. 3. add gender to question options
